Add new workspace with `.RData` files
New example workspace `workspace-rmd-RData_iris-mtcars` contains the `iris` and `mtcars` data.frames, which are included in `R`, saved as one `example.RData` file.
Workspace further contains an `RMarkdown` script file, which reads the `.RData` file and creates simple plots with the data included.
Thus it demonstrates the platforms ability to find, read and work with `.RData` files and `data.frames`.